+++ Browser/TreeCategoryModel.cpp
@@ -1,4 +1,7 @@
-// SPDX-FileCopyrightText: 2003-2022 The KPhotoAlbum Development Team
+// SPDX-FileCopyrightText: 2009 Jan Kundrát <jkt@flaska.net>
+// SPDX-FileCopyrightText: 2009-2022 Jesper K. Pedersen <jesper.pedersen@kdab.com>
+// SPDX-FileCopyrightText: 2013-2023 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
+// SPDX-FileCopyrightText: 2015-2022 Tobias Leupold <tl@stonemx.de>
 //
 // SPDX-License-Identifier: GPL-2.0-or-later
 
@@ -209,7 +212,7 @@ bool Browser::TreeCategoryModel::dropMimeData(const QMimeData *data, Qt::DropAct
     if (action == Qt::MoveAction) {
         // Remove the tag from its group and remove the group if it's empty now
         m_memberMap.removeMemberFromGroup(m_category->name(), droppedTagData.tagGroup, droppedTagData.tagName);
-        if (m_memberMap.members(m_category->name(), droppedTagData.tagGroup, true) == QStringList()) {
+        if (m_memberMap.members(m_category->name(), droppedTagData.tagGroup, true).isEmpty()) {
             m_memberMap.deleteGroup(m_category->name(), droppedTagData.tagGroup);
         }
     }